+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 2.39.0 (Unreleased)
-- Added support for LSYOI (thanks [@zymooll](https://github.com/zymooll))
+- Added support for LSYOI and SDUT OnlineJudge (thanks [@zymooll](https://github.com/zymooll) for the LSYOI parser)
 
 ## [2.38.2](https://github.com/jmerle/competitive-companion/releases/tag/2.38.2) (2022-12-24)
 - Fixed the parsing of Kattis problems